requirements.txt
asgiref==3.5.2
Django==4.1
django-ckeditor==6.5.1
django-ckeditor-5==0.1.9
django-debug-toolbar==3.6.0
django-js-asset==2.0.0
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2022.2
